Move Adapters to a separate package
