+++ app/src/main/java/com/disarm/sanna/pdm/Adapters/MyAdapter.java
@@ -1,4 +1,4 @@
-package com.disarm.sanna.pdm;
+package com.disarm.sanna.pdm.Adapters;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -6,7 +6,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import java.util.List;
+import com.disarm.sanna.pdm.R;
 
 /**
  * Created by Sanna on 20-06-2016.
