+++ app/src/main/java/com/disarm/sanna/pdm/Adapters/MyAdapterActivityList.java
@@ -1,4 +1,4 @@
-package com.disarm.sanna.pdm;
+package com.disarm.sanna.pdm.Adapters;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -6,6 +6,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.disarm.sanna.pdm.R;
 
 /**
  * Created by Sanna on 22-06-2016.
