+++ app/src/main/java/com/disarm/sanna/pdm/MainActivity.java
@@ -30,6 +30,7 @@ import android.view.View;
 import android.widget.CompoundButton;
 import android.widget.Toast;
 
+import com.disarm.sanna.pdm.Adapters.MyAdapter;
 import com.disarm.sanna.pdm.DisarmConnect.MyService;
 import com.disarm.sanna.pdm.Service.SyncService;
 import com.disarm.sanna.pdm.Util.DividerItemDecoration;
